// src/modules/auth/dto/auth-response.dto/auth-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';
export class AuthResponseDto {
  @ApiProperty()
  accessToken: string;
  @ApiProperty()
  refreshToken: string;
  @ApiProperty()
  user: {
    id: string;
    email: string;
    firstName: string;
    lastName: string;
    role: string;
    psn?: string;
  };
}